requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pillow==10.2.0
yarl==1.9.4
